fix: make functions to validate registration fields
