--- src/shared/lib/validation/address.ts
@@ -1,9 +1,0 @@
-import { object, string } from "yup";
-import { matchOnlyAlphabeticCharacters } from "./helpers";
-
-export const addressSchema = object({
-  country: string().required("Выберите страну"),
-  city: matchOnlyAlphabeticCharacters(string().required("Введите город")),
-  streetName: string().required("Введите улицу"),
-  postalCode: string().required("Введите почтовый индекс"),
-});
