+++ src/shared/lib/validation/index.ts
@@ -1,0 +1,8 @@
+export {
+  matchOnlyAlphabeticCharacters,
+  validateName,
+  validateEmail,
+  validatePassword,
+  validatePasswordConfirm,
+  validateDate,
+} from "./registration";
